requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py38_0
cffi=1.14.0=py38he30daa8_1
chardet=3.0.4=py38_1003
cryptography=2.9.2=py38h1ba5d50_0
cudatoolkit=10.0.130=0
cycler=0.10.0=py_2
dbus=1.13.14=hb2f20db_0
expat=2.2.9=he1b5a44_2
fontconfig=2.13.1=he4413a7_1000
freetype=2.9.1=h8a8886c_1
glib=2.63.1=h3eb4bd4_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
icu=58.2=hf484d3e_1000
idna=2.9=py_1
intel-openmp=2020.1=217
joblib=0.15.1=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.2.0=py38hbf85e49_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.3=he6710b0_1
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1006
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=2.32.1=h14c3975_1000
libxcb=1.13=h14c3975_1002
libxml2=2.9.9=h13577e0_2
matplotlib=3.1.3=py38_0
matplotlib-base=3.1.3=py38hef1b27d_0
mkl=2020.1=217
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.0.15=py38ha843d7b_0
mkl_random=1.1.1=py38h0573a6f_0
ncurses=6.2=he6710b0_1
ninja=1.9.0=py38hfd86e86_0
numpy=1.18.1=py38h4f9e942_0
numpy-base=1.18.1=py38hde5b4d6_1
olefile=0.46=py_0
openssl=1.1.1g=h7b6447c_0
pandas=1.0.3=py38h0573a6f_0
pcre=8.44=he1b5a44_0
pillow=7.1.2=py38hb39fc2d_0
pip=20.0.2=py38_3
pthread-stubs=0.4=h14c3975_1001
pycparser=2.20=py_0
pyopenssl=19.1.0=py38_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py38h05f1152_4
pysocks=1.7.1=py38_0
python=3.8.3=hcff3b4d_0
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.4.0=py3.8_cuda10.0.130_cudnn7.6.3_0
pytz=2020.1=py_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
requests=2.23.0=py38_0
scikit-learn=0.23.1=pypi_0
scipy=1.4.1=pypi_0
setuptools=46.4.0=py38_0
signatory=1.2.1.1.4.0=pypi_0
sip=4.19.13=py38he6710b0_0
six=1.14.0=py38_0
sqlite=3.31.1=h62c20be_1
threadpoolctl=2.1.0=pypi_0
tk=8.6.8=hbc83047_0
torchvision=0.5.0=py38_cu100
tornado=6.0.4=py38h1e0a361_1
tqdm=4.46.0=pyh9f0ad1d_0
urllib3=1.25.8=py38_0
wheel=0.34.2=py38_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
